Refactoring: moved TagSettings into its own component
