+++ src/apps/project-settings/TagSettings/index.ts
@@ -1,0 +1,1 @@
+export * from './TagSettings';
